Fixes compiler warning: integer precision
10 warnings in PlayerEntityStickProfile.m
Changed some arguments and return values from int to NSUInteger:
jostick axis, control index, point count

+++ src/Core/OOJoystickProfile.h
@@ -78,9 +78,9 @@ MA 02110-1301, USA.
 - (id) init;
 - (void) dealloc;
 - (id) copyWithZone: (NSZone *) zone;
-- (int) addControl: (NSPoint) point;
+- (NSInteger) addControl: (NSPoint) point;
 - (NSPoint) pointAtIndex: (NSInteger) index;
-- (int) countPoints;
+- (NSInteger) countPoints;
 - (void) removeControl: (NSInteger) index;
 - (void) clearControlPoints;
 - (void) moveControl: (NSInteger) index point: (NSPoint) point;
